backend/requirements.txt
bcrypt
fastapi
uvicorn
motor
pydantic
python-dotenv
email-validator
python-jose[cryptography]
python-multipart